Add DW_OP_-prefixed env vars available for database
This allows for having more-secure passwords for databases that have that requirement, such as they have if they ran the secure setup script that comes with a Homebrew-installed MySQL.

If absent, the values fall back to the ones already in place.

In order to document the presence of these env vars, I changed the docker-compose.yml's used environment variables.

I tried to set a wrong database, and got a good 500 error in the development mode of the website.

To add more environment variables, we can re-use this prefix. Perhaps it becomes easy enough to search the codebase for it, to find and document.